admin/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
jinja2==3.1.4
python-multipart==0.0.9
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
google-cloud-firestore==2.14.0
